#123 - Optimize app for production Remove unneccesary modules and imports Optimized about page assets Removed mock data Fix linting issues
