--- functions/index.js
@@ -1,4 +1,0 @@
-const functions = require('firebase-functions');
-
-const admin = require('firebase-admin');
-admin.initializeApp();
