+++ package.json
@@ -1,37 +1,35 @@
 {
-  "name": "food-hub",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "start": "npm run serve"
-  },
-  "dependencies": {
-    "axios": "^0.19.2",
-    "core-js": "^3.6.4",
-    "firebase": "^7.14.1",
-    "firebase-admin": "^8.10.0",
-    "firebase-functions": "^3.6.0",
-    "vue": "^2.6.11",
-    "vue-carousel": "^0.18.0",
-    "vue-headful": "^2.1.0",
-    "vue-loading-overlay": "^3.3.2",
-    "vue-router": "^3.1.5",
-    "vuejs-paginate": "^2.1.0",
-    "vuex": "^3.1.2"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.2.0",
-    "@vue/cli-plugin-eslint": "^4.2.0",
-    "@vue/cli-service": "^4.2.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "babel-eslint": "^10.0.3",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-vue": "^6.1.2",
-    "prettier": "^1.19.1",
-    "vue-template-compiler": "^2.6.11"
-  }
+	"name": "food-hub",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"serve": "vue-cli-service serve",
+		"build": "vue-cli-service build",
+		"lint": "vue-cli-service lint",
+		"start": "npm run serve"
+	},
+	"dependencies": {
+		"axios": "^0.19.2",
+		"core-js": "^3.6.4",
+		"firebase": "^7.14.1",
+		"vue": "^2.6.11",
+		"vue-carousel": "^0.18.0",
+		"vue-headful": "^2.1.0",
+		"vue-loading-overlay": "^3.3.2",
+		"vue-router": "^3.1.5",
+		"vuejs-paginate": "^2.1.0",
+		"vuex": "^3.1.2"
+	},
+	"devDependencies": {
+		"@vue/cli-plugin-babel": "^4.2.0",
+		"@vue/cli-plugin-eslint": "^4.2.0",
+		"@vue/cli-service": "^4.2.0",
+		"@vue/eslint-config-prettier": "^6.0.0",
+		"babel-eslint": "^10.0.3",
+		"eslint": "^6.7.2",
+		"eslint-plugin-prettier": "^3.1.1",
+		"eslint-plugin-vue": "^6.1.2",
+		"prettier": "^1.19.1",
+		"vue-template-compiler": "^2.6.11"
+	}
 }
