+++ public/index.html
@@ -1,30 +1,31 @@
 <!DOCTYPE html>
 <html lang="en">
+	<head>
+		<meta charset="utf-8" />
+		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
+		<meta name="viewport" content="width=device-width,initial-scale=1.0" />
+		<link
+			href="https://fonts.googleapis.com/css2?family=Poppins&display=swap"
+			rel="stylesheet"
+		/>
+		<link
+			rel="stylesheet"
+			href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css"
+		/>
+		<link rel="icon" href="<%= BASE_URL %>favicon.ico" />
+		<title>Food Hub</title>
+	</head>
 
-<head>
-	<meta charset="utf-8" />
-	<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-	<meta name="viewport" content="width=device-width,initial-scale=1.0" />
-	<link href="https://fonts.googleapis.com/css2?family=Poppins&display=swap" rel="stylesheet" />
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.13.0/css/all.min.css" />
-	<link rel="icon" href="<%= BASE_URL %>favicon.ico" />
-	<title>Food Hub</title>
-
-</head>
-
-<body>
-	<noscript>
-		<strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't
-			work properly without JavaScript enabled. Please enable it to
-			continue.</strong>
-	</noscript>
-	<div id="app"></div>
-	<script src="/__/firebase/7.14.1/firebase-app.js"></script>
-	<script src="/__/firebase/7.14.1/firebase-analytics.js"></script>
-	<script src="/__/firebase/7.14.1/firebase-auth.js"></script>
-	<script src="/__/firebase/7.14.1/firebase-firestore.js"></script>
-</body>
-
+	<body>
+		<noscript>
+			<strong
+				>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't
+				work properly without JavaScript enabled. Please enable it to
+				continue.</strong
+			>
+		</noscript>
+		<div id="app"></div>
+	</body>
 </html>
 
 <style scoped>
