+++ src/components/carousel/Carousel.vue
