+++ src/components/recipe/Ingredients.vue
