+++ src/components/recipe/NutritionFacts.vue
@@ -101,7 +101,7 @@
 		</div>
 	</div>
 </template>
- 
+
 <script>
 export default {
 	props: {
