--- src/mock/RecipeData.js
@@ -1,1541 +1,0 @@
-export function getMockRecipeList() {
-	return [
-		{
-			id: 90629,
-			image: "https://spoonacular.com/recipeImages/90629-312x231.jpg",
-			title: "Baked Apples in White Wine",
-		},
-		{
-			id: 284420,
-			image: "https://spoonacular.com/recipeImages/284420-312x231.jpg",
-			title: "Chocolate Silk Pie with Marshmallow Meringue",
-		},
-		{
-			id: 73420,
-			image: "https://spoonacular.com/recipeImages/73420-312x231.jpg",
-			title: "Apple Or Peach Strudel",
-		},
-		{
-			id: 632660,
-			image: "https://spoonacular.com/recipeImages/632660-312x231.jpg",
-			title: "Apricot Glazed Apple Tart",
-		},
-		{
-			id: 716429,
-			calories: 584,
-			carbs: "84g",
-			fat: "20g",
-			image: "https://spoonacular.com/recipeImages/716429-312x231.jpg",
-			imageType: "jpg",
-			protein: "19g",
-			title: "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs",
-		},
-		{
-			id: 715538,
-			calories: 521,
-			carbs: "69g",
-			fat: "10g",
-			image: "https://spoonacular.com/recipeImages/715538-312x231.jpg",
-			imageType: "jpg",
-			protein: "35g",
-			title:
-				"What to make for dinner tonight?? Bruschetta Style Pork & Pasta",
-		},
-		{
-			id: 90629,
-			image: "https://spoonacular.com/recipeImages/90629-312x231.jpg",
-			title: "Baked Apples in White Wine",
-		},
-		{
-			id: 284420,
-			image: "https://spoonacular.com/recipeImages/284420-312x231.jpg",
-			title: "Chocolate Silk Pie with Marshmallow Meringue",
-		},
-		{
-			id: 73420,
-			image: "https://spoonacular.com/recipeImages/73420-312x231.jpg",
-			title: "Apple Or Peach Strudel",
-		},
-		{
-			id: 632660,
-			image: "https://spoonacular.com/recipeImages/632660-312x231.jpg",
-			title: "Apricot Glazed Apple Tart",
-		},
-		{
-			id: 716429,
-			calories: 584,
-			carbs: "84g",
-			fat: "20g",
-			image: "https://spoonacular.com/recipeImages/716429-312x231.jpg",
-			imageType: "jpg",
-			protein: "19g",
-			title: "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs",
-		},
-		{
-			id: 715538,
-			calories: 521,
-			carbs: "69g",
-			fat: "10g",
-			image: "https://spoonacular.com/recipeImages/715538-312x231.jpg",
-			imageType: "jpg",
-			protein: "35g",
-			title:
-				"What to make for dinner tonight?? Bruschetta Style Pork & Pasta",
-		},
-		{
-			id: 90629,
-			image: "https://spoonacular.com/recipeImages/90629-312x231.jpg",
-			title: "Baked Apples in White Wine",
-		},
-		{
-			id: 284420,
-			image: "https://spoonacular.com/recipeImages/284420-312x231.jpg",
-			title: "Chocolate Silk Pie with Marshmallow Meringue",
-		},
-		{
-			id: 73420,
-			image: "https://spoonacular.com/recipeImages/73420-312x231.jpg",
-			title: "Apple Or Peach Strudel",
-		},
-		{
-			id: 632660,
-			image: "https://spoonacular.com/recipeImages/632660-312x231.jpg",
-			title: "Apricot Glazed Apple Tart",
-		},
-	];
-}
-
-export function getRecipeById(id) {
-    switch (id) {
-        case '90629':
-            return {
-                "preparationMinutes": 15,
-                "cookingMinutes": 80,
-                "extendedIngredients": [
-                    {
-                        "original": "1/4 cup apple juice",
-                    },
-                    {
-                        "original": "2 tablespoons firmly-packed light brown sugar",
-                    },
-                    {
-                        "original": "1 tablespoon butter",
-                    },
-                    {
-                        "original": "1/4 teaspoon cinnamon",
-                    },
-                    {
-                        "original": "1/2 cup dry white wine",
-                    },
-                    {
-                        "original": "1/3 cup sultana raisin",
-                    },
-                    {
-                        "original": "4 red delicious apples",
-                    }
-                ],
-                "id": 90629,
-                "title": "Baked Apples in White Wine",
-                "readyInMinutes": 95,
-                "servings": 4,
-                "image": "https://spoonacular.com/recipeImages/90629-556x370.jpg",
-                "nutrition": {
-                    "nutrients": [
-                        {
-                            "title": "Calories",
-                            "amount": 210.34,
-                            "unit": "cal",
-                            "percentOfDailyNeeds": 10.52
-                        },
-                        {
-                            "title": "Fat",
-                            "amount": 3.24,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 4.98
-                        },
-                        {
-                            "title": "Saturated Fat",
-                            "amount": 1.87,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 11.71
-                        },
-                        {
-                            "title": "Carbohydrates",
-                            "amount": 43.14,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 14.38
-                        },
-                        {
-                            "title": "Net Carbohydrates",
-                            "amount": 37.85,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 13.76
-                        },
-                        {
-                            "title": "Sugar",
-                            "amount": 26.51,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 29.46
-                        },
-                        {
-                            "title": "Cholesterol",
-                            "amount": 7.53,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 2.51
-                        },
-                        {
-                            "title": "Sodium",
-                            "amount": 34.01,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 1.48
-                        },
-                        {
-                            "title": "Alcohol",
-                            "amount": 3.09,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 17.17
-                        },
-                        {
-                            "title": "Protein",
-                            "amount": 0.86,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 1.71
-                        },
-                        {
-                            "title": "Fiber",
-                            "amount": 5.29,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 21.15
-                        },
-                        {
-                            "title": "Vitamin C",
-                            "amount": 9.17,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 11.11
-                        },
-                        {
-                            "title": "Potassium",
-                            "amount": 340.74,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 9.74
-                        },
-                        {
-                            "title": "Manganese",
-                            "amount": 0.17,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 8.41
-                        },
-                        {
-                            "title": "Vitamin B6",
-                            "amount": 0.12,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 5.89
-                        },
-                        {
-                            "title": "Copper",
-                            "amount": 0.09,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.6
-                        },
-                        {
-                            "title": "Vitamin B2",
-                            "amount": 0.08,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.57
-                        },
-                        {
-                            "title": "Magnesium",
-                            "amount": 17.18,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.3
-                        }
-                    ]
-                },
-                "summary": "Baked Apples in White Wine might be just the dessert you are searching for. One portion of this dish contains about <b>1g of protein</b>, <b>3g of fat</b>, and a total of <b>210 calories</b>. This recipe serves 4. For <b>$1.37 per serving</b>, this recipe <b>covers 4%</b> of your daily requirements of vitamins and minerals. 1 person were glad they tried this recipe. It is a good option if you're following a <b>gluten free and vegetarian</b> diet. If you have cinnamon, wine, tasty apples, and a few other ingredients on hand, you can make it. From preparation to the plate, this recipe takes roughly <b>1 hour and 35 minutes</b>. All things considered, we decided this recipe <b>deserves a spoonacular score of 13%</b>. This score is not so great. Try <a href=\"https://spoonacular.com/recipes/a-simple-yet-elegant-dessert-baked-apples-in-red-wine-515386\">A Simple Yet Elegant Dessert – Baked Apples In Red Wine</a>, <a href=\"https://spoonacular.com/recipes/oven-baked-fish-in-white-wine-91152\">Oven Baked Fish in White Wine</a>, and <a href=\"https://spoonacular.com/recipes/chicken-baked-in-white-wine-marinade-340382\">Chicken Baked in White Wine Marinade</a> for similar recipes.",
-                "analyzedInstructions": []
-            };
-        case '284420':
-            return {
-                "preparationMinutes": 15,
-                "cookingMinutes": 135,
-                "extendedIngredients": [
-                    {
-                        "original": "5 oz. BAKER'S Bittersweet Chocolate, divided"
-                    },
-                    {
-                        "original": "4 oz. (1/2 of 8-oz. pkg.) PHILADELPHIA Cream Cheese, softened"
-                    },
-                    {
-                        "original": "1 pkg. (3.9 oz.) JELL-O Chocolate Flavor Instant Pudding"
-                    },
-                    {
-                        "original": "1 jar (7 oz.) JET-PUFFED Marshmallow Creme, divided"
-                    },
-                    {
-                        "original": "1 cup cold milk"
-                    },
-                    {
-                        "original": "1 OREO Pie Crust (6 oz.)"
-                    },
-                    {
-                        "original": "1 cup thawed COOL WHIP Whipped Topping"
-                    }
-                ],
-                "id": 284420,
-                "title": "Chocolate Silk Pie with Marshmallow Meringue",
-                "readyInMinutes": 135,
-                "servings": 15,
-                "image": "https://spoonacular.com/recipeImages/284420-556x370.jpg",
-                "nutrition": {
-                    "nutrients": [
-                        {
-                            "title": "Calories",
-                            "amount": 225.87,
-                            "unit": "cal",
-                            "percentOfDailyNeeds": 11.29
-                        },
-                        {
-                            "title": "Fat",
-                            "amount": 10.03,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 15.44
-                        },
-                        {
-                            "title": "Saturated Fat",
-                            "amount": 5.04,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 31.47
-                        },
-                        {
-                            "title": "Carbohydrates",
-                            "amount": 32.62,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 10.87
-                        },
-                        {
-                            "title": "Net Carbohydrates",
-                            "amount": 31.4,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 11.42
-                        },
-                        {
-                            "title": "Sugar",
-                            "amount": 22.11,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 24.56
-                        },
-                        {
-                            "title": "Cholesterol",
-                            "amount": 10.61,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.54
-                        },
-                        {
-                            "title": "Sodium",
-                            "amount": 198.63,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 8.64
-                        },
-                        {
-                            "title": "Caffeine",
-                            "amount": 8.64,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 2.88
-                        },
-                        {
-                            "title": "Protein",
-                            "amount": 2.44,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 4.87
-                        },
-                        {
-                            "title": "Manganese",
-                            "amount": 0.15,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 7.69
-                        },
-                        {
-                            "title": "Copper",
-                            "amount": 0.15,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 7.5
-                        },
-                        {
-                            "title": "Phosphorus",
-                            "amount": 70.9,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 7.09
-                        },
-                        {
-                            "title": "Magnesium",
-                            "amount": 25.89,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 6.47
-                        },
-                        {
-                            "title": "Iron",
-                            "amount": 1.06,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 5.9
-                        },
-                        {
-                            "title": "Fiber",
-                            "amount": 1.23,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 4.9
-                        },
-                        {
-                            "title": "Vitamin B2",
-                            "amount": 0.07,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.25
-                        },
-                        {
-                            "title": "Calcium",
-                            "amount": 36.08,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.61
-                        },
-                        {
-                            "title": "Potassium",
-                            "amount": 121.88,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.48
-                        },
-                        {
-                            "title": "Vitamin B1",
-                            "amount": 0.05,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.11
-                        },
-                        {
-                            "title": "Vitamin A",
-                            "amount": 136.31,
-                            "unit": "IU",
-                            "percentOfDailyNeeds": 2.73
-                        },
-                        {
-                            "title": "Selenium",
-                            "amount": 1.88,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 2.69
-                        },
-                        {
-                            "title": "Zinc",
-                            "amount": 0.4,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 2.67
-                        },
-                        {
-                            "title": "Folate",
-                            "amount": 9.95,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 2.49
-                        },
-                        {
-                            "title": "Vitamin B3",
-                            "amount": 0.4,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 2.02
-                        },
-                        {
-                            "title": "Vitamin B12",
-                            "amount": 0.12,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 1.99
-                        },
-                        {
-                            "title": "Vitamin D",
-                            "amount": 0.26,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 1.71
-                        },
-                        {
-                            "title": "Vitamin B5",
-                            "amount": 0.13,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 1.34
-                        },
-                        {
-                            "title": "Vitamin K",
-                            "amount": 1.22,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 1.16
-                        }
-                    ]
-                },
-                "summary": "This recipe serves 15 and costs 70 cents per serving. One portion of this dish contains around <b>2g of protein</b>, <b>10g of fat</b>, and a total of <b>226 calories</b>. 1 person has made this recipe and would make it again. A mixture of marshmallow creme, philadelphia cream cheese, milk, and a handful of other ingredients are all it takes to make this recipe so scrumptious. From preparation to the plate, this recipe takes around <b>2 hours and 15 minutes</b>. All things considered, we decided this recipe <b>deserves a spoonacular score of 12%</b>. This score is not so spectacular. Try <a href=\"https://spoonacular.com/recipes/chocolate-silk-pie-with-marshmallow-meringue-135373\">Chocolate Silk Pie With Marshmallow Meringue</a>, <a href=\"https://spoonacular.com/recipes/homemade-french-silk-pie-from-scratch-909019\">Homemade French Silk Pie (From Scratch)</a>, and <a href=\"https://spoonacular.com/recipes/coconut-cake-with-silk-meringue-buttercream-562702\">Coconut Cake with Silk Meringue Buttercream</a> for similar recipes.",
-                "analyzedInstructions": [
-                    {
-                        "name": "",
-                        "steps": [
-                            {
-                                "number": 1,
-                                "step": "Microwave 4 oz. chocolate in medium microwaveable bowl on HIGH 1 to 1-1/2 min. or until melted, stirring after 1 min.",
-                            },
-                            {
-                                "number": 2,
-                                "step": "Add cream cheese, half the marshmallow creme, dry pudding mix and milk; beat with mixer until blended."
-                            },
-                            {
-                                "number": 3,
-                                "step": "Pour into crust."
-                            },
-                            {
-                                "number": 4,
-                                "step": "Beat remaining marshmallow creme with COOL WHIP in separate bowl until blended; spread over chocolate layer in crust."
-                            },
-                            {
-                                "number": 5,
-                                "step": "Refrigerate 2 hours or until firm.  Meanwhile, use remaining chocolate to make chocolate curls.  Refrigerate until ready to use."
-                            },
-                            {
-                                "number": 6,
-                                "step": "Garnish pie with chocolate curls before serving."
-                            }
-                        ]
-                    }
-                ]
-            };
-        case '73420':
-            return {
-                "extendedIngredients": [
-                    {
-                        "original": "6 large baking apples",
-                    },
-                    {
-                        "original": "1 tsp baking powder",
-                    },
-                    {
-                        "original": "1 tsp cinnamon",
-                    },
-                    {
-                        "original": "1 egg",
-                    },
-                    {
-                        "original": "1 cup flour",
-                    },
-                    {
-                        "original": "1/2 tsp salt",
-                    },
-                    {
-                        "original": "1 cup sugar",
-                    }
-                ],
-                "id": 73420,
-                "title": "Apple Or Peach Strudel",
-                "readyInMinutes": 16,
-                "servings": 8,
-                "image": "https://spoonacular.com/recipeImages/73420-556x370.jpg",
-                "nutrition": {
-                    "nutrients": [
-                        {
-                            "title": "Calories",
-                            "amount": 249.87,
-                            "unit": "cal",
-                            "percentOfDailyNeeds": 12.49
-                        },
-                        {
-                            "title": "Fat",
-                            "amount": 0.97,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 1.49
-                        },
-                        {
-                            "title": "Saturated Fat",
-                            "amount": 0.24,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 1.53
-                        },
-                        {
-                            "title": "Carbohydrates",
-                            "amount": 60.61,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 20.2
-                        },
-                        {
-                            "title": "Net Carbohydrates",
-                            "amount": 55.99,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 20.36
-                        },
-                        {
-                            "title": "Sugar",
-                            "amount": 42.4,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 47.11
-                        },
-                        {
-                            "title": "Cholesterol",
-                            "amount": 20.46,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 6.82
-                        },
-                        {
-                            "title": "Sodium",
-                            "amount": 155.98,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 6.78
-                        },
-                        {
-                            "title": "Protein",
-                            "amount": 2.75,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 5.51
-                        },
-                        {
-                            "title": "Fiber",
-                            "amount": 4.62,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 18.49
-                        },
-                        {
-                            "title": "Manganese",
-                            "amount": 0.23,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 11.37
-                        },
-                        {
-                            "title": "Vitamin B1",
-                            "amount": 0.15,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 10.22
-                        },
-                        {
-                            "title": "Selenium",
-                            "amount": 7.15,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 10.21
-                        },
-                        {
-                            "title": "Vitamin C",
-                            "amount": 7.71,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 9.34
-                        },
-                        {
-                            "title": "Folate",
-                            "amount": 36.22,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 9.05
-                        },
-                        {
-                            "title": "Phosphorus",
-                            "amount": 89.3,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 8.93
-                        },
-                        {
-                            "title": "Vitamin B2",
-                            "amount": 0.15,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 8.86
-                        },
-                        {
-                            "title": "Potassium",
-                            "amount": 268.32,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 7.67
-                        },
-                        {
-                            "title": "Iron",
-                            "amount": 1.11,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 6.19
-                        },
-                        {
-                            "title": "Vitamin B3",
-                            "amount": 1.08,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 5.42
-                        },
-                        {
-                            "title": "Calcium",
-                            "amount": 46.13,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.61
-                        },
-                        {
-                            "title": "Vitamin B6",
-                            "amount": 0.09,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.27
-                        },
-                        {
-                            "title": "Copper",
-                            "amount": 0.07,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.74
-                        },
-                        {
-                            "title": "Vitamin K",
-                            "amount": 3.84,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 3.66
-                        },
-                        {
-                            "title": "Magnesium",
-                            "amount": 12.84,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.21
-                        },
-                        {
-                            "title": "Vitamin B5",
-                            "amount": 0.26,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 2.56
-                        },
-                        {
-                            "title": "Vitamin E",
-                            "amount": 0.38,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 2.5
-                        },
-                        {
-                            "title": "Vitamin A",
-                            "amount": 120.97,
-                            "unit": "IU",
-                            "percentOfDailyNeeds": 2.42
-                        },
-                        {
-                            "title": "Zinc",
-                            "amount": 0.26,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 1.74
-                        }
-                    ],
-                },
-                "summary": "Apple Or Peach Strudel might be just the dessert you are searching for. This recipe makes 8 servings with <b>234 calories</b>, <b>3g of protein</b>, and <b>1g of fat</b> each. For <b>56 cents per serving</b>, this recipe <b>covers 5%</b> of your daily requirements of vitamins and minerals. 1 person has made this recipe and would make it again. It is a good option if you're following a <b>dairy free and vegetarian</b> diet. From preparation to the plate, this recipe takes about <b>16 minutes</b>. Head to the store and pick up baking apples, baking powder, salt, and a few other things to make it today. All things considered, we decided this recipe <b>deserves a spoonacular score of 27%</b>. This score is not so outstanding. Try <a href=\"https://spoonacular.com/recipes/apple-strudel-apfel-strudel-73414\">Apple Strudel (apfel Strudel)</a>, <a href=\"https://spoonacular.com/recipes/peach-strudel-507356\">Peach Strudel</a>, and <a href=\"https://spoonacular.com/recipes/peach-walnut-strudel-149300\">Peach Walnut Strudel</a> for similar recipes.",
-                "analyzedInstructions": [],
-            };
-        case '632660':
-            return {
-                "extendedIngredients": [
-                    {
-                        "original": "1 1/2 cups all-purpose flour, plus 1 tablespoon",
-                    },
-                    {
-                        "original": "Pinch of salt",
-                    },
-                    {
-                        "original": "1 1/2 sticks cold unsalted butter cold unsalted butter<",
-                    },
-                    {
-                        "original": "cup ice water",
-                    },
-                    {
-                        "original": "3 1/2 tablespoons sugar",
-                    },
-                    {
-                        "original": "4 larges red apples, such as Golden Delicious, peeled, cored and cut into 1/4-inch-thick slices",
-                    },
-                    {
-                        "original": "2 teaspoons cinnamon",
-                    },
-                    {
-                        "original": "2 tablespoons apricot preserves, melted and strained",
-                    }
-                ],
-                "id": 632660,
-                "title": "Apricot Glazed Apple Tart",
-                "readyInMinutes": 45,
-                "servings": 4,
-                "image": "https://spoonacular.com/recipeImages/632660-556x370.jpg",
-                "nutrition": {
-                    "nutrients": [
-                        {
-                            "title": "Calories",
-                            "amount": 657.72,
-                            "unit": "cal",
-                            "percentOfDailyNeeds": 32.89
-                        },
-                        {
-                            "title": "Fat",
-                            "amount": 35.24,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 54.22
-                        },
-                        {
-                            "title": "Saturated Fat",
-                            "amount": 21.91,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 136.92
-                        },
-                        {
-                            "title": "Carbohydrates",
-                            "amount": 84.34,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 28.11
-                        },
-                        {
-                            "title": "Net Carbohydrates",
-                            "amount": 77.16,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 28.06
-                        },
-                        {
-                            "title": "Sugar",
-                            "amount": 38.16,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 42.4
-                        },
-                        {
-                            "title": "Cholesterol",
-                            "amount": 91.11,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 30.37
-                        },
-                        {
-                            "title": "Sodium",
-                            "amount": 322.58,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 14.03
-                        },
-                        {
-                            "title": "Protein",
-                            "amount": 5.89,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 11.78
-                        },
-                        {
-                            "title": "Manganese",
-                            "amount": 0.58,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 28.84
-                        },
-                        {
-                            "title": "Fiber",
-                            "amount": 7.18,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 28.71
-                        },
-                        {
-                            "title": "Vitamin B1",
-                            "amount": 0.41,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 27.21
-                        },
-                        {
-                            "title": "Vitamin A",
-                            "amount": 1202.82,
-                            "unit": "IU",
-                            "percentOfDailyNeeds": 24.06
-                        },
-                        {
-                            "title": "Selenium",
-                            "amount": 16.61,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 23.73
-                        },
-                        {
-                            "title": "Folate",
-                            "amount": 93.9,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 23.48
-                        },
-                        {
-                            "title": "Vitamin B2",
-                            "amount": 0.31,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 18.15
-                        },
-                        {
-                            "title": "Vitamin B3",
-                            "amount": 3.01,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 15.03
-                        },
-                        {
-                            "title": "Iron",
-                            "amount": 2.59,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 14.38
-                        },
-                        {
-                            "title": "Vitamin C",
-                            "amount": 11.18,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 13.55
-                        },
-                        {
-                            "title": "Vitamin E",
-                            "amount": 1.45,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 9.66
-                        },
-                        {
-                            "title": "Potassium",
-                            "amount": 311.16,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 8.89
-                        },
-                        {
-                            "title": "Phosphorus",
-                            "amount": 86.27,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 8.63
-                        },
-                        {
-                            "title": "Vitamin K",
-                            "amount": 8.32,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 7.93
-                        },
-                        {
-                            "title": "Copper",
-                            "amount": 0.15,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 7.57
-                        },
-                        {
-                            "title": "Magnesium",
-                            "amount": 23.9,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 5.98
-                        },
-                        {
-                            "title": "Vitamin B6",
-                            "amount": 0.12,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 5.85
-                        },
-                        {
-                            "title": "Calcium",
-                            "amount": 44.49,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 4.45
-                        },
-                        {
-                            "title": "Vitamin D",
-                            "amount": 0.64,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 4.24
-                        },
-                        {
-                            "title": "Vitamin B5",
-                            "amount": 0.39,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.94
-                        },
-                        {
-                            "title": "Zinc",
-                            "amount": 0.49,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 3.25
-                        },
-                        {
-                            "title": "Vitamin B12",
-                            "amount": 0.07,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 1.2
-                        }
-                    ],
-                },
-                "summary": "If you have about <b>about 45 minutes</b> to spend in the kitchen, Apricot Glazed Apple Tart might be a great <b>lacto ovo vegetarian</b> recipe to try. For <b>$1.59 per serving</b>, you get a dessert that serves 4. One serving contains <b>658 calories</b>, <b>6g of protein</b>, and <b>35g of fat</b>. It is brought to you by Foodista. Not a lot of people made this recipe, and 3 would say it hit the spot. If you have cinnamon, pinch of salt, butter butter, and a few other ingredients on hand, you can make it. Overall, this recipe earns a <b>not so great spoonacular score of 29%</b>. If you like this recipe, take a look at these similar recipes: <a href=\"https://spoonacular.com/recipes/tarte-aux-abricots-glazed-french-apricot-tart-with-almonds-124725\">Tarte Aux Abricots - Glazed French Apricot Tart With Almonds</a>, <a href=\"https://spoonacular.com/recipes/apricot-apple-glazed-gammon-513531\">Apricot & Apple Glazed Gammon</a>, and <a href=\"https://spoonacular.com/recipes/orange-glazed-apple-apricot-crisp-481367\">Orange glazed Apple Apricot Crisp</a>.",
-                "analyzedInstructions": [
-                    {
-                        "steps": [
-                            {
-                                "number": 1,
-                                "step": "In a food processor, pulse 1 1/2 cups of the flour with the salt.",
-                            },
-                            {
-                                "number": 2,
-                                "step": "Add the cold butter and process just until the butter is the size of peas, about 5 seconds. Sprinkle the ice water over the mixture and process just until moistened, about 5 seconds.",
-                            },
-                            {
-                                "number": 3,
-                                "step": "Transfer the dough to a lightly floured work surface and knead 2 or 3 times, just until it comes together. Pat the dough into a disk. On a lightly floured work surface, roll out the dough into a 16- to 17-inch round about 1/4 inch thick.",
-                            },
-                            {
-                                "number": 4,
-                                "step": "Line a large baking sheet with parchment paper.",
-                            },
-                            {
-                                "number": 5,
-                                "step": "Roll the dough around the rolling pin and unroll it onto the prepared baking sheet.",
-                            },
-                            {
-                                "number": 6,
-                                "step": "In a small bowl, combine 2 tablespoons of the sugar with the remaining 1 tablespoon of flour and sprinkle over the dough. Arrange the apple slices on top in overlapping concentric circles to within 3 inches of the edge. Fold the dough over the apples in a free-form fashion.",
-                            },
-                            {
-                                "number": 7,
-                                "step": "Brush the apples with the melted butter and sprinkle with the remaining 1 1/2 tablespoons of sugar and cinnamon. Refrigerate the unbaked tart until slightly chilled, about 10 minutes.",
-                            },
-                            {
-                                "number": 8,
-                                "step": "Preheat the oven to 40",
-                            },
-                            {
-                                "number": 9,
-                                "step": "Bake the tart in the center of the oven for 1 hour, or until the apples are tender and golden and the crust is deep golden and cooked through.",
-                            },
-                            {
-                                "number": 10,
-                                "step": "Brush the apples with the melted preserves. Slide the parchment onto a wire rack and let the tart cool slightly before serving.",
-                            }
-                        ]
-                    }
-                ],
-            };
-        case '716429':
-            return {
-                "extendedIngredients": [
-                    {
-                        "original": "1 tbsp butter",
-                    },
-                    {
-                        "original": "about 2 cups frozen cauliflower florets, thawed, cut into bite-sized pieces",
-                    },
-                    {
-                        "original": "2 tbsp grated cheese (I used romano)",
-                    },
-                    {
-                        "original": "1-2 tbsp extra virgin olive oil",
-                    },
-                    {
-                        "original": "5-6 cloves garlic",
-                    },
-                    {
-                        "original": "6-8 ounces pasta (I used linguine)",
-                    },
-                    {
-                        "original": "couple of pinches red pepper flakes, optional",
-                    },
-                    {
-                        "original": "salt and pepper, to taste",
-                    },
-                    {
-                        "original": "3 scallions, chopped, white and green parts separated",
-                    },
-                    {
-                        "original": "2-3 tbsp white wine",
-                    },
-                    {
-                        "original": "1/4 cup whole wheat bread crumbs (I used panko)",
-                    }
-                ],
-                "id": 716429,
-                "title": "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs",
-                "readyInMinutes": 45,
-                "servings": 2,
-                "image": "https://spoonacular.com/recipeImages/716429-556x370.jpg",
-                "nutrition": {
-                    "nutrients": [
-                        {
-                            "title": "Calories",
-                            "amount": 584.46,
-                            "unit": "cal",
-                            "percentOfDailyNeeds": 29.22
-                        },
-                        {
-                            "title": "Fat",
-                            "amount": 19.83,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 30.51
-                        },
-                        {
-                            "title": "Saturated Fat",
-                            "amount": 8.09,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 50.57
-                        },
-                        {
-                            "title": "Carbohydrates",
-                            "amount": 83.71,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 27.9
-                        },
-                        {
-                            "title": "Net Carbohydrates",
-                            "amount": 76.98,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 27.99
-                        },
-                        {
-                            "title": "Sugar",
-                            "amount": 5.36,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 5.95
-                        },
-                        {
-                            "title": "Cholesterol",
-                            "amount": 31.02,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 10.34
-                        },
-                        {
-                            "title": "Sodium",
-                            "amount": 451.42,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 19.63
-                        },
-                        {
-                            "title": "Alcohol",
-                            "amount": 1.55,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 8.58
-                        },
-                        {
-                            "title": "Protein",
-                            "amount": 18.98,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 37.97
-                        },
-                        {
-                            "title": "Selenium",
-                            "amount": 57.72,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 82.45
-                        },
-                        {
-                            "title": "Vitamin C",
-                            "amount": 56.62,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 68.64
-                        },
-                        {
-                            "title": "Manganese",
-                            "amount": 1.11,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 55.52
-                        },
-                        {
-                            "title": "Vitamin K",
-                            "amount": 58.27,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 55.49
-                        },
-                        {
-                            "title": "Phosphorus",
-                            "amount": 304.38,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 30.44
-                        },
-                        {
-                            "title": "Fiber",
-                            "amount": 6.73,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 26.92
-                        },
-                        {
-                            "title": "Folate",
-                            "amount": 87.14,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 21.79
-                        },
-                        {
-                            "title": "Vitamin B6",
-                            "amount": 0.43,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 21.46
-                        },
-                        {
-                            "title": "Magnesium",
-                            "amount": 71.55,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 17.89
-                        },
-                        {
-                            "title": "Calcium",
-                            "amount": 178.12,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 17.81
-                        },
-                        {
-                            "title": "Potassium",
-                            "amount": 597.53,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 17.07
-                        },
-                        {
-                            "title": "Iron",
-                            "amount": 3.07,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 17.04
-                        },
-                        {
-                            "title": "Copper",
-                            "amount": 0.33,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 16.43
-                        },
-                        {
-                            "title": "Zinc",
-                            "amount": 2.12,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 14.15
-                        },
-                        {
-                            "title": "Vitamin B5",
-                            "amount": 1.17,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 11.69
-                        },
-                        {
-                            "title": "Vitamin B2",
-                            "amount": 0.2,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 11.5
-                        },
-                        {
-                            "title": "Vitamin A",
-                            "amount": 537.51,
-                            "unit": "IU",
-                            "percentOfDailyNeeds": 10.75
-                        },
-                        {
-                            "title": "Vitamin B3",
-                            "amount": 2.14,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 10.71
-                        },
-                        {
-                            "title": "Vitamin B1",
-                            "amount": 0.16,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 10.46
-                        },
-                        {
-                            "title": "Vitamin E",
-                            "amount": 1.53,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 10.2
-                        },
-                        {
-                            "title": "Vitamin B12",
-                            "amount": 0.14,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 2.28
-                        },
-                        {
-                            "title": "Vitamin D",
-                            "amount": 0.2,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 1.31
-                        }
-                    ],
-                },
-                "summary": "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs might be just the main course you are searching for. This recipe makes 2 servings with <b>636 calories</b>, <b>21g of protein</b>, and <b>20g of fat</b> each. For <b>$1.83 per serving</b>, this recipe <b>covers 24%</b> of your daily requirements of vitamins and minerals. From preparation to the plate, this recipe takes about <b>45 minutes</b>. This recipe is liked by 209 foodies and cooks. If you have pasta, salt and pepper, cheese, and a few other ingredients on hand, you can make it. To use up the extra virgin olive oil you could follow this main course with the <a href=\"https://spoonacular.com/recipes/peach-crisp-healthy-crisp-for-breakfast-487698\">Peach Crisp: Healthy Crisp for Breakfast</a> as a dessert. All things considered, we decided this recipe <b>deserves a spoonacular score of 86%</b>. This score is tremendous. Try <a href=\"https://spoonacular.com/recipes/cauliflower-gratin-with-garlic-breadcrumbs-318375\">Cauliflower Gratin with Garlic Breadcrumbs</a>, <a href=\"https://spoonacular.com/recipes/pasta-with-cauliflower-sausage-breadcrumbs-30437\">Pasta With Cauliflower, Sausage, & Breadcrumbs</a>, and <a href=\"https://spoonacular.com/recipes/pasta-with-roasted-cauliflower-parsley-and-breadcrumbs-30738\">Pasta With Roasted Cauliflower, Parsley, And Breadcrumbs</a> for similar recipes.",
-                "analyzedInstructions": [],
-            };
-        case '715538':
-            return {
-                "preparationMinutes": 5,
-                "cookingMinutes": 30,
-                "extendedIngredients": [
-                    {
-                        "original": "3 cups bow tie pasta",
-                    },
-                    {
-                        "original": "½ cup Parmigiano Reggiano",
-                    },
-                    {
-                        "original": "Kraft Recipe Makers Chicken Bruschetta Pasta",
-                    },
-                    {
-                        "original": "1-1/2 lb. pork chops",
-                    }
-                ],
-                "id": 715538,
-                "title": "What to make for dinner tonight?? Bruschetta Style Pork & Pasta",
-                "readyInMinutes": 35,
-                "servings": 5,
-                "image": "https://spoonacular.com/recipeImages/715538-556x370.jpg",
-                "nutrition": {
-                    "nutrients": [
-                        {
-                            "title": "Calories",
-                            "amount": 521.13,
-                            "unit": "cal",
-                            "percentOfDailyNeeds": 26.06
-                        },
-                        {
-                            "title": "Fat",
-                            "amount": 10.27,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 15.8
-                        },
-                        {
-                            "title": "Saturated Fat",
-                            "amount": 4.12,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 25.74
-                        },
-                        {
-                            "title": "Carbohydrates",
-                            "amount": 69.02,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 23.01
-                        },
-                        {
-                            "title": "Net Carbohydrates",
-                            "amount": 66.07,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 24.03
-                        },
-                        {
-                            "title": "Sugar",
-                            "amount": 2.54,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 2.82
-                        },
-                        {
-                            "title": "Cholesterol",
-                            "amount": 67.58,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 22.53
-                        },
-                        {
-                            "title": "Sodium",
-                            "amount": 209.26,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 9.1
-                        },
-                        {
-                            "title": "Protein",
-                            "amount": 35.12,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 70.24
-                        },
-                        {
-                            "title": "Selenium",
-                            "amount": 90.42,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 129.17
-                        },
-                        {
-                            "title": "Vitamin B1",
-                            "amount": 0.69,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 46.12
-                        },
-                        {
-                            "title": "Phosphorus",
-                            "amount": 448.3,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 44.83
-                        },
-                        {
-                            "title": "Vitamin B3",
-                            "amount": 8.84,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 44.19
-                        },
-                        {
-                            "title": "Manganese",
-                            "amount": 0.85,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 42.6
-                        },
-                        {
-                            "title": "Vitamin B6",
-                            "amount": 0.8,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 39.92
-                        },
-                        {
-                            "title": "Zinc",
-                            "amount": 2.98,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 19.86
-                        },
-                        {
-                            "title": "Magnesium",
-                            "amount": 76.75,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 19.19
-                        },
-                        {
-                            "title": "Copper",
-                            "amount": 0.32,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 15.99
-                        },
-                        {
-                            "title": "Potassium",
-                            "amount": 552.74,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 15.79
-                        },
-                        {
-                            "title": "Vitamin B2",
-                            "amount": 0.26,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 15.07
-                        },
-                        {
-                            "title": "Calcium",
-                            "amount": 144.07,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 14.41
-                        },
-                        {
-                            "title": "Fiber",
-                            "amount": 2.94,
-                            "unit": "g",
-                            "percentOfDailyNeeds": 11.78
-                        },
-                        {
-                            "title": "Vitamin B5",
-                            "amount": 1.1,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 11.01
-                        },
-                        {
-                            "title": "Vitamin B12",
-                            "amount": 0.6,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 10.01
-                        },
-                        {
-                            "title": "Iron",
-                            "amount": 1.73,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 9.62
-                        },
-                        {
-                            "title": "Folate",
-                            "amount": 17.26,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 4.31
-                        },
-                        {
-                            "title": "Vitamin D",
-                            "amount": 0.41,
-                            "unit": "µg",
-                            "percentOfDailyNeeds": 2.75
-                        },
-                        {
-                            "title": "Vitamin A",
-                            "amount": 81.73,
-                            "unit": "IU",
-                            "percentOfDailyNeeds": 1.63
-                        },
-                        {
-                            "title": "Vitamin E",
-                            "amount": 0.24,
-                            "unit": "mg",
-                            "percentOfDailyNeeds": 1.61
-                        }
-                    ],
-                },
-                "summary": "You can never have too many main course recipes, so give What to make for dinner tonight?? Bruschetta Style Pork & Pastan a try. For <b>$1.45 per serving</b>, this recipe <b>covers 22%</b> of your daily requirements of vitamins and minerals. This recipe makes 5 servings with <b>521 calories</b>, <b>35g of protein</b>, and <b>10g of fat</b> each. 163 people have tried and liked this recipe. If you have bow tie pasta, pork chops, recipe makers chicken bruschetta pasta, and a few other ingredients on hand, you can make it. To use up the pasta you could follow this main course with the <a href=\"https://spoonacular.com/recipes/chocolate-dessert-pasta-138646\">Chocolate Dessert Pasta</a> as a dessert. This recipe is typical of Mediterranean cuisine. From preparation to the plate, this recipe takes around <b>35 minutes</b>. All things considered, we decided this recipe <b>deserves a spoonacular score of 90%</b>. This score is awesome. Try <a href=\"https://spoonacular.com/recipes/dinner-tonight-grilled-romesco-style-pork-209128\">Dinner Tonight: Grilled Romesco-Style Pork</a>, <a href=\"https://spoonacular.com/recipes/dinner-tonight-chickpea-bruschetta-31868\">Dinner Tonight: Chickpea Bruschetta</a>, and <a href=\"https://spoonacular.com/recipes/dinner-tonight-pasta-with-pork-loin-chinese-eggplant-baby-bok-choy-and-spicy-miso-sauce-206706\">Dinner Tonight: Pasta with Pork Loin, Chinese Eggplant, Baby Bok Choy, and Spicy Miso Sauce</a> for similar recipes.",
-                "analyzedInstructions": [
-                    {
-                        "steps": [
-                            {
-                                "number": 1,
-                                "step": "wash and rinse pork chops and place into the skillet.cut them into bite sized pieces and add half of the Basil Garlic simmer sauce.boil your water and start working on cooking your bow-tie pasta.when you have finished with boiling and draining your pasta, add it to the pork along with the rest of the Basil Garlic Simmering Sauce, mixing lightly.Next you will top with the Chunky Bruschetta Finishing Sauce, cover with Parmesan, and cover. Cooking on low heat 2 to 3 minutes or until heated through.",
-                            }
-                        ]
-                    }
-                ],
-            }
-    }
-}
-
-export function getMockSearchRecipesComplex() {
-    return {
-        offset: 0,
-        number: 4,
-        results: [
-            {
-                id: 716429,
-                calories: 584,
-                carbs: "84g",
-                fat: "20g",
-                image:
-                    "https://spoonacular.com/recipeImages/716429-312x231.jpg",
-                imageType: "jpg",
-                protein: "19g",
-                title:
-                    "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs",
-            },
-            {
-                id: 715538,
-                calories: 521,
-                carbs: "69g",
-                fat: "10g",
-                image:
-                    "https://spoonacular.com/recipeImages/715538-312x231.jpg",
-                imageType: "jpg",
-                protein: "35g",
-                title:
-                    "What to make for dinner tonight?? Bruschetta Style Pork & Pasta",
-            },
-            {
-                id: 716429,
-                calories: 584,
-                carbs: "84g",
-                fat: "20g",
-                image:
-                    "https://spoonacular.com/recipeImages/716429-312x231.jpg",
-                imageType: "jpg",
-                protein: "19g",
-                title:
-                    "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs",
-            },
-            {
-                id: 715538,
-                calories: 521,
-                carbs: "69g",
-                fat: "10g",
-                image:
-                    "https://spoonacular.com/recipeImages/715538-312x231.jpg",
-                imageType: "jpg",
-                protein: "35g",
-                title:
-                    "What to make for dinner tonight?? Bruschetta Style Pork & Pasta",
-            },
-        ],
-        totalResults: 86,
-    };
-}
-
-export function getMockResultsByName() {
-    return {
-        offset: 0,
-        number: 4,
-        results: [
-            {
-                id: 633508,
-                image:
-                    "https://spoonacular.com/recipeImages/633508-312x231.jpg",
-                imageUrls: ["Baked-Cheese-Manicotti-633508.jpg"],
-                readyInMinutes: 45,
-                servings: 6,
-                title: "Baked Cheese Manicotti",
-            },
-            {
-                id: 634873,
-                image:
-                    "https://spoonacular.com/recipeImages/634873-312x231.jpg",
-                imageUrls: ["Best-Baked-Macaroni-and-Cheese-634873.jpg"],
-                readyInMinutes: 45,
-                servings: 12,
-                title: "Best Baked Macaroni and Cheese",
-            },
-            {
-                id: 633508,
-                image:
-                    "https://spoonacular.com/recipeImages/633508-312x231.jpg",
-                imageUrls: ["Baked-Cheese-Manicotti-633508.jpg"],
-                readyInMinutes: 45,
-                servings: 6,
-                title: "Baked Cheese Manicotti",
-            },
-            {
-                id: 634873,
-                image:
-                    "https://spoonacular.com/recipeImages/634873-312x231.jpg",
-                imageUrls: ["Best-Baked-Macaroni-and-Cheese-634873.jpg"],
-                readyInMinutes: 45,
-                servings: 12,
-                title: "Best Baked Macaroni and Cheese",
-            },
-        ],
-        totalResults: 719,
-    };
-}
