+++ src/router/index.js
@@ -72,7 +72,7 @@ const routes = [
 		path: "/about",
 		name: "About",
 		component: About,
-	}
+	},
 ];
 
 function authorizeRouting(to, from, next) {
